rename files to meet PEP8
see https://www.python.org/dev/peps/pep-0008/#package-and-module-names
