+++ dtos/v1/word_cloud_dto_v1.py
