+++ models/__init__.py
@@ -1,5 +1,5 @@
-from models.AccountModel import Account
-from models.DataModel import Data
-from models.PredictionModel import Prediction
-from models.VisitorModel import Visitor
-from models.QAModel import QAModel
+from models.account_model import Account
+from models.data_model import Data
+from models.prediction_model import Prediction
+from models.visitor_model import Visitor
+from models.qa_model import QAModel
