+++ models/prediction_model.py
@@ -15,15 +15,15 @@ class Prediction(db.Model):
     # For declaring foreign key : db.Foreignkey('tablename.columnname')
     # For specifying the relationship, db.relationship('packagename.filename.classname', back_populates = 'columnname specified in the class')
     model_id = db.Column(Integer, db.ForeignKey('qa_model.id'), nullable=False)
-    model = db.relationship('models.QAModel.QAModel', back_populates='predictions')
+    model = db.relationship('models.qa_model.QAModel', back_populates='predictions')
 
     # many to one relationship with Data
     datum_id = db.Column(Integer, db.ForeignKey('data.id'), nullable=False)
-    datum = db.relationship('models.DataModel.Data', back_populates='predictions')
+    datum = db.relationship('models.data_model.Data', back_populates='predictions')
 
     # many to one relationship with Model
     visitor_id = db.Column(Integer, db.ForeignKey('visitor.id'), nullable=False)
-    visitor = db.relationship('models.VisitorModel.Visitor', back_populates='predictions')
+    visitor = db.relationship('models.visitor_model.Visitor', back_populates='predictions')
 
     def __repr__(self) -> str:
         return (f'Prediction {self.id} '+ f'Prediction:{self.prediction} ' + f'Alt_ans:{self.alt_answer}')
