+++ models/qa_model.py
@@ -6,7 +6,7 @@ class QAModel(db.Model):
     __tablename__ = 'qa_model'
 
     id = db.Column(Integer, primary_key=True, autoincrement = True)
-    
+
     ml_type = db.Column(String(100), nullable=False)
     ml_version = db.Column(String(20), nullable=False)
 
@@ -17,7 +17,7 @@ class QAModel(db.Model):
     created_date = db.Column(DateTime, nullable=False)
 
     # One to many relationship with prediction
-    predictions = db.relationship('models.PredictionModel.Prediction', back_populates='model')
+    predictions = db.relationship('models.prediction_model.Prediction', back_populates='model')
 
     # tostring code use for testing and debug
     def __repr__(self) -> str:
